+++ src/main/java/efub/team4/backend_eweather/domain/bear/controller/BearController.java
@@ -2,6 +2,7 @@ package efub.team4.backend_eweather.domain.bear.controller;
 
 import efub.team4.backend_eweather.domain.bear.dto.BearImageResponseDto;
 import efub.team4.backend_eweather.domain.bear.service.BearService;
+import efub.team4.backend_eweather.domain.weather.service.OpenWeatherAPI;
 import lombok.RequiredArgsConstructor;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,9 +20,12 @@ public class BearController {
     @Autowired
     private final BearService bearService;
 
+    @Autowired
+    private final OpenWeatherAPI openWeatherAPI;
+
     @GetMapping
     public BearImageResponseDto loadBearImage() throws IOException, ParseException {
-        return bearService.findBearImage();
+        return bearService.findBearImage(openWeatherAPI.findBearInfo());
     }
 
 }
