+++ src/main/java/efub/team4/backend_eweather/domain/weather/service/OpenWeatherAPI.java
@@ -1,7 +1,6 @@
 package efub.team4.backend_eweather.domain.weather.service;
 
 import efub.team4.backend_eweather.domain.weather.dto.*;
-import efub.team4.backend_eweather.global.util.TimeUtil;
 import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.json.simple.parser.JSONParser;
@@ -23,11 +22,10 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
-
+@Service
 @RequiredArgsConstructor
 public class OpenWeatherAPI {
 
-    TimeUtil timeUtil = new TimeUtil();
 
     // 커밋 시 서비스 키 지우고 커밋
     private final String BASE_URL = "http://apis.data.go.kr/1360000/VilageFcstInfoService_2.0/getVilageFcst";
@@ -36,7 +34,7 @@ public class OpenWeatherAPI {
     private String pageNo = "1";
     private String numOfRows = "310";
     private String dataType = "JSON";
-    private String baseTime = timeUtil.getBaseTime(); // api 제공시각
+    private String baseTime = "0500"; // api 제공시각
     private String nx = "59";
     private String ny = "126"; // nx, ny는 서대문구 신촌동 좌표값
 
@@ -105,7 +103,7 @@ public class OpenWeatherAPI {
 
     public URL buildRequestUrl() throws IOException {
         StringBuilder sb = new StringBuilder(BASE_URL);
-        String baseDate = timeUtil.getBaseTime();
+        String baseDate = getCurrentDate();
         sb.append("?").append(URLEncoder.encode("serviceKey", "UTF-8")).append("=").append(serviceKey);
         sb.append("&").append(URLEncoder.encode("pageNo", "UTF-8")).append("=").append(URLEncoder.encode(pageNo, "UTF-8"));
         sb.append("&").append(URLEncoder.encode("numOfRows", "UTF-8")).append("=").append(URLEncoder.encode(numOfRows, "UTF-8")); /* 한 페이지 결과 수 */
@@ -176,7 +174,7 @@ public class OpenWeatherAPI {
             obj = (JSONObject) parseItem.get(idx+5);
             sky = (String) obj.get("fcstValue");
             ForcastResponseDto dto = ForcastResponseDto.builder()
-                    .baseDate(timeUtil.getCurrentDate())
+                    .baseDate(getCurrentDate())
                     .baseTime(baseTime)
                     .fcstDate(fcstDate)
                     .fcstTime(fcstTime)
@@ -198,11 +196,11 @@ public class OpenWeatherAPI {
 
     private CalendarWeatherResponseDto buildCalendarData(JSONArray jsonResult) {
         // 현재 시간 받아서 예상 시간 조회
-        String fcstTime = timeUtil.getFcstTime();
+        String fcstTime = getFcstTime();
         if(fcstTime.length() == 3){
             fcstTime = "0" + fcstTime;
         }
-        String baseDate = timeUtil.getCurrentDate();
+        String baseDate = getCurrentDate();
         String tmp = "";
         String tmx = "";
         String tmn = "";
@@ -267,11 +265,11 @@ public class OpenWeatherAPI {
 
     public WeatherResponseDto builldWeatherResponse(JSONArray jsonResult, String category) {
         // 현재 시간 받아서 예상 시간 조회
-        String fcstTime = timeUtil.getFcstTime();
+        String fcstTime = getFcstTime();
         if (fcstTime.length() == 3) {
             fcstTime = "0" + fcstTime;
         }
-        String baseDate = timeUtil.getCurrentDate();
+        String baseDate = getCurrentDate();
         String value = "";
 
         // jsonResult를 조회하며 필요한 데이터 받기
@@ -332,11 +330,11 @@ public class OpenWeatherAPI {
 
     private BearResponseDto buildBearResponse(JSONArray jsonResult) {
 
-        String fcstTime = timeUtil.getFcstTime();
+        String fcstTime = getFcstTime();
         if (fcstTime.length() == 3) {
             fcstTime = "0" + fcstTime;
         }
-        String baseDate = timeUtil.getCurrentDate();
+        String baseDate = getCurrentDate();
         String tmp = "";
         String tmx = "";
         String tmn = "";
@@ -377,6 +375,57 @@ public class OpenWeatherAPI {
                 .build();
 
         return responseDto;
+    }
+
+    public String getCurrentDate(){
+        DateFormat simpleDateFormat = new SimpleDateFormat("yyyyMMdd");
+        Date currentDate = new Date();
+        return simpleDateFormat.format(currentDate);
+    }
+
+    public String getFcstTime(){
+        DateFormat simpleDateFormat = new SimpleDateFormat("k");
+        Date currentTime = new Date();
+        return simpleDateFormat.format(currentTime) + "00";
+    }
+
+    /*
+    public String getBaseTime(){
+        DateFormat simpleDateFormat = new SimpleDateFormat("k");
+        Date currentHour = new Date();
+        int ApiTime = Integer.parseInt(String.valueOf(currentHour));
+        String baseTime = switch (ApiTime) {
+            case 0, 1, 2 -> "2300";
+            case 3, 4, 5 -> "0200";
+            case 6, 7, 8 -> "0500";
+            case 9, 10, 11 -> "0800";
+            case 12, 13, 14 -> "1100";
+            case 15, 16, 17 -> "1400";
+            case 18, 19, 20 -> "1700";
+            case 21, 22, 23 -> "2000";
+            default -> "";
+        };
+        return baseTime;
+    }
+
+     */
+
+    public String getDay(){
+        DateFormat simpleDateFormat = new SimpleDateFormat("k");
+        Date currentTime = new Date();
+        Integer time = Integer.parseInt(simpleDateFormat.format(currentTime));
+        if(time > 6 && time < 18 ){
+            return "day";
+        }
+        else{
+            return "night";
+        }
+    }
+
+    public Integer getMonth(){
+        DateFormat simpleDateFormat = new SimpleDateFormat("M");
+        Date currentTime = new Date();
+        return Integer.parseInt(simpleDateFormat.format(currentTime));
     }
 
 }
